requirements.txt
matplotlib==3.10.0
numpy==2.2.2
pandas==2.2.3
Pillow==11.1.0
tensorflow==2.18.0
tensorflow_intel==2.18.0
torch==2.5.1+cu121
torchmetrics==1.6.1
torchvision==0.20.1+cu121
tqdm==4.67.1
